:wrench: chore: Colocando validações no timevalidator de hora
